+++ src/middleware/notFound.ts
@@ -1,8 +1,0 @@
-import type { MiddlewareCreator } from "./lib/common.js";
-
-const notFound: MiddlewareCreator = () => async (_req, res) => {
-  res.statusCode = 404;
-  res.end();
-};
-
-export default notFound;