+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [Unreleased]
 ### Changed
 - fix: make getBuilder optional #41 
+### Added
+- feat: Expose as middleware #46
 
 ## [0.10.0] - 2023-05-22
 ### Added